Move expression and statement terms to lib
